+++ plataforma-academia/app [academia]/academia-application/src/main/java/br/com/academia/application/repository/EnderecoRepository.java
@@ -3,7 +3,8 @@ package br.com.academia.application.repository;
 import br.com.academia.application.model.Endereco;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface EnderecoRepository extends JpaRepository<Endereco, Long> {
+
+public interface EnderecoRepository extends JpaRepository<Endereco, Integer> {
 
 
 }
